# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  modules.AliDNS:
    properties:
      comment:
        type: string
      content:
        type: string
      id:
        type: string
      name:
        type: string
      priority:
        description: Data     string `json:"data"`
        type: integer
      ttl:
        type: integer
      type:
        type: string
    type: object
  web.Domain:
    properties:
      data: {}
      errors:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: makiras.x@outlook.com
    name: domain0
  description: Domain0 API
  license:
    name: MPL(mozilla public license)-2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/
  title: Domain0 API
  version: 0.0.1
paths:
...
  /api/v1/domain/{id}/dns:
    get:
      consumes:
      - application/json
      description: |-
        List Domain Dns **AliDNS as Example, read modules for others**
        user must have read permission to domain or be admin
      parameters:
      - description: domain id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/modules.AliDNS'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
      summary: List Domain Dns
      tags:
      - domain
    post:
      consumes:
      - application/json
      description: |-
        Create Domain Dns **AliDNS as Example, read modules for others**
        user must have readwrite permission to domain or be admin
        for now only owner can edit domain which ICP_reg is true
      parameters:
      - description: domain id
        in: path
        name: id
        required: true
        type: string
      - description: dns info
        in: body
        name: dns
        required: true
        schema:
          $ref: '#/definitions/modules.AliDNS'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  $ref: '#/definitions/modules.AliDNS'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
      summary: Create Domain Dns
      tags:
      - domain
  /api/v1/domain/{id}/dns/{dnsId}:
    delete:
      consumes:
      - application/json
      description: |-
        Delete Domain Dns
        user must have readwrite permission to domain or be admin
      parameters:
      - description: domain id
        in: path
        name: id
        required: true
        type: string
      - description: dns id
        in: path
        name: dnsId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
      summary: Delete Domain Dns
      tags:
      - domain
    put:
      consumes:
      - application/json
      description: |-
        Update Domain Dns **AliDNS as Example, read modules for others**
        user must have readwrite permission to domain or be admin
        for now only owner can edit domain which ICP_reg is true
      parameters:
      - description: domain id
        in: path
        name: id
        required: true
        type: string
      - description: dns id
        in: path
        name: dnsId
        required: true
        type: string
      - description: dns info
        in: body
        name: dns
        required: true
        schema:
          $ref: '#/definitions/modules.AliDNS'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  $ref: '#/definitions/modules.AliDNS'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
      summary: Update Domain Dns
      tags:
      - domain
...
schemes:
- http
- https
swagger: "2.0"